% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gibber.R
\name{create_gibber}
\alias{create_gibber}
\title{A function for random word generation.}
\usage{
create_gibber(n, prob = NULL)
}
\arguments{
\item{n}{a numeric vector giving the number of characters in each word.}

\item{prob}{a vector of probability weights for obtaining the elements of the vector being sampled.}
}
\value{
a character vector.
}
\description{
Compute a word of length n consisting of randomly chosen
characters from the alphabet.
}
\examples{
# generate a random word of length 5
create_gibber(n=5)
# generate three randoms word:
create_gibber(c(5, 7, 9))
}
