% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibber.R
\docType{package}
\name{gibber}
\alias{gibber}
\title{gibber: a package for detecting gibberish text.}
\description{
A recurrent neural network was trained to recognize gibber in text. For this purpose,
dataset was used that contains 700.000 English words (50\% gibberish words/50\% natural words).
Special care was taken to avoid overfitting and an accuracy of 97\% was achieved on a validation
dataset, thereby guaranteeing generalizable results.
}
\section{For a quick start}{

\code{\link[gibber]{is_gibber}} \cr
}

