% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_gibber.R
\name{is_gibber}
\alias{is_gibber}
\title{Detect presence of gibberish text in a sentence}
\usage{
is_gibber(sentence, output = "prob", trim_sentence = 2)
}
\arguments{
\item{sentence}{character vector. A sentence or words.}

\item{output}{either "prob" for probabilities, "class" for classifications or "bool" for TRUE/FALSE ("prob" by default).}

\item{trim_sentence}{integer. Specify how many characters a word in a sentence must have to be excluded from the prediction model (2 by default).}
}
\value{
numeric vector
}
\description{
Assess the probability that a sentence contains gibberish words. For each word in a sentence, the model
inspects the sequence of vowels and consonents to estimate whether a sentence consists of natural words. Therefore,
words like 'asdfg' and 'dfrgfh' are considered unnatural and are classified as not being part of the English spoken language.
}
\details{
A recurrent neural network was trained for this purpose, using a dataset that contains 700.000 words (50\% gibberish words/50\% natural words).
Special care was taken to avoid overfitting and an accuracy of 97\% was achieved on a validation dataset, thereby guaranteeing generalizable results.
}
\examples{
sentence <- c("asdf", "This is great", "computer")
is_gibber(sentence)
}
