% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize_words.R
\name{tokenize_words}
\alias{tokenize_words}
\title{Tensorize words on the character level.}
\usage{
tokenize_words(words)
}
\arguments{
\item{words}{a character vector.}
}
\value{
a matrix.
}
\description{
All inputted words are transformed and put into a matrix with two dimensions.
Each row represents a word. Each column refers to a positional letter from the encyphered word,
where the last column maps to the first letter of the word and the second last column to the second letter of the words, etc.
Each cell refers to a letter from the alphabet. For example, a cell with the number one refers to the letter 'e'.
}
\details{
A word can't surpass the length of 31 characters and only letters from the alphabet are contained, others characters like numbers and aren't encyphered.
}
\examples{
words <- c("telephone", "laptop")
tokenize_words(words)
}
