% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_gibber.R
\name{is_gibber}
\alias{is_gibber}
\title{Predict presence of gibberish words in a sentence}
\usage{
is_gibber(words)
}
\arguments{
\item{words}{character vector}
}
\value{

}
\description{
A recurrent neural network is used to calculate the probability of
}
\examples{
words <- c("asdf", "elephant")
is_gibber(words)
}
