% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_gibber.R
\name{is_gibber}
\alias{is_gibber}
\title{Detect presence of gibberish text in a sentence}
\usage{
is_gibber(sentence, output = "prob", trim_sentence = 2)
}
\arguments{
\item{sentence}{character vector. A sentence or words.}

\item{output}{either "prob" for probabilities, "class" for classifications or "bool" for TRUE/FALSE ("prob" by default).}

\item{trim_sentence}{integer. Specify how many characters a word in a sentence must have to be excluded from the prediction model (2 by default).}
}
\value{
numeric vector
}
\description{
A recurrent neural network is used to calculate the...
}
\examples{
sentence <- c("asdf", "This is great", "computer")
is_gibber(sentence)
}
